extract EthosComponents.tsx
