+++ components/index.ts
@@ -28,3 +28,4 @@ export * from "./SiteLayout"
 export * from "./TwitterSection"
 export * from "./WipItem"
 export * from "./WipList"
+export * from "./EthosComponents"
